+++ citydb-config/src/main/java/module-info.java
@@ -2,7 +2,6 @@ module org.citydb.config {
     requires com.alibaba.fastjson2;
 
     exports org.citydb.config;
-    exports org.citydb.config.configs;
 
     opens org.citydb.config to com.alibaba.fastjson2;
 }
