+++ citydb-config/src/main/java/org/citydb/config/Config.java
@@ -23,71 +23,25 @@ package org.citydb.config;
 
 import com.alibaba.fastjson2.JSON;
 import com.alibaba.fastjson2.JSONObject;
-import com.alibaba.fastjson2.JSONWriter;
-import com.alibaba.fastjson2.annotation.JSONField;
-import org.citydb.config.configs.ExportConfig;
-import org.citydb.config.configs.ImportConfig;
 
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Objects;
 
 public class Config {
-    @JSONField(name = "import")
-    private ImportConfig importConfig;
-    @JSONField(name = "export")
-    private ExportConfig exportConfig;
-    @JSONField(name = "plugins")
-    private Map<String, JSONObject> pluginsConfig;
+    private Map<String, JSONObject> configs = new HashMap<>();
 
-    public ImportConfig getImportConfig() {
-        if (importConfig == null) {
-            importConfig = new ImportConfig();
+    public void setConfigs(Map<String, JSONObject> configs) {
+        if (configs != null) {
+            this.configs = configs;
         }
-
-        return importConfig;
     }
 
-    public Config setImportConfig(ImportConfig importConfig) {
-        this.importConfig = importConfig;
-        return this;
-    }
-
-    public ExportConfig getExportConfig() {
-        if (exportConfig == null) {
-            exportConfig = new ExportConfig();
-        }
-
-        return exportConfig;
-    }
-
-    public Config setExportConfig(ExportConfig exportConfig) {
-        this.exportConfig = exportConfig;
-        return this;
-    }
-
-    public Map<String, JSONObject> getPluginsConfig() {
-        if (pluginsConfig == null) {
-            pluginsConfig = new HashMap<>();
-        }
-
-        return pluginsConfig;
-    }
-
-    public Config setPluginsConfig(Map<String, JSONObject> pluginsConfig) {
-        this.pluginsConfig = pluginsConfig;
-        return this;
-    }
-
-    public void writeTo(Path target) throws ConfigException {
-        Objects.requireNonNull(target, "The target path must not be null.");
-        try {
-            Files.write(target, JSON.toJSONString(this, JSONWriter.Feature.PrettyFormat).getBytes());
-        } catch (IOException e) {
-            throw new ConfigException("Failed to write config file to " + target, e);
+    public <T> T getConfig(String name, Class<T> type) throws ConfigException {
+        JSONObject options = configs.get(name);
+        if (options != null) {
+            return JSON.parseObject(options.toString(), type);
+        } else {
+            throw new ConfigException("Unsupported configuration type '" + type.getName() + "'.");
         }
     }
 }
